+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.0 under development
 
 - Chg #221: Throw `LogicException` in `Tag::id()` when id is empty string (@razvbir)
+- Chg #267: Make all `CheckboxItem` and `RadioItem` properties required (@vjik)
 
 ## 3.13.0 March 13, 2026